Added templates
Instead of using standard HttpResponse, I rendered templates that were more complecated. I created a templates folder in the main mysite directory and made two simple templates for the homepage. To fix a TemplateDoesNotExistError, I added the path to the settings file. I also fixed a problem with the views file being incorrectly called in the url file.