Added more complicated views and urls for the polls app
I created a template for the polls homepage that shows a list of all the questions. Additionally, if you type a question id, eg. polls/1, it will the question text. Also if you cleck on the question text from the homepage it will send you to the details page. Finally, I added in 404 error if the question you are looking for does not exist.